IFIX-313: Department entity create api

+++ domain-services/ifix-department-entity-service/src/main/java/org/egov/web/models/DepartmentEntity.java
@@ -40,7 +40,7 @@ public class DepartmentEntity {
     private String name = null;
 
     @JsonProperty("hierarchyLevelId")
-    private BigDecimal hierarchyLevelId = null;
+    private Integer hierarchyLevelId = null;
 
     @JsonProperty("children")
     @Valid
